pyproject.toml
[project]
name = "robotflow-connectors"
version = "0.1.0"
description = "Secure multi-provider LLM connectors for robotics — Claude, Codex, MiniMax, GLM-5, Kimi with OAuth and API key auth"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AIFLOW LABS LIMITED", email = "ilessio@aiflowlabs.io" },
]
keywords = [
    "llm",
    "robotics",
    "anthropic",
    "openai",
    "claude",
    "codex",
    "minimax",
    "streaming",
    "oauth",
    "robotflow",
    "reachy",
    "ai-agent",
    "multi-provider",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "vcrpy>=6.0",
    "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/RobotFlow-Labs/robotflow-connector"
Repository = "https://github.com/RobotFlow-Labs/robotflow-connector"
Issues = "https://github.com/RobotFlow-Labs/robotflow-connector/issues"
Documentation = "https://github.com/RobotFlow-Labs/robotflow-connector#readme"

[project.scripts]
robotflow-connectors = "robotflow_connectors.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/robotflow_connectors"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]
